Moved bulk of WebServer stuff to new WebServer class
