+++ include/hub.h
@@ -2,6 +2,7 @@
 #define HUB_H
 
 #include <Arduino.h>
+#include <WebServer.h>
 #include "config.h"
 #include "controller.h"
 
@@ -11,7 +12,7 @@ public:
   void loop() override;
   void handleClient(WiFiClient& client);
 private:
-  WiFiServer server;
+  WebServer http;
 };
 
 #endif
